Package structure: moved UI utils
Part of #963.
