+++ src/main/java/de/dennisguse/opentracks/chart/ChartView.java
@@ -49,9 +49,9 @@ import de.dennisguse.opentracks.data.models.Marker;
 import de.dennisguse.opentracks.stats.ExtremityMonitor;
 import de.dennisguse.opentracks.ui.markers.MarkerDetailActivity;
 import de.dennisguse.opentracks.ui.markers.MarkerUtils;
+import de.dennisguse.opentracks.ui.util.ThemeUtils;
 import de.dennisguse.opentracks.util.IntentUtils;
 import de.dennisguse.opentracks.util.StringUtils;
-import de.dennisguse.opentracks.util.ThemeUtils;
 
 /**
  * Visualization of the chart.
